[UPDATE] Refine classes for the texture

+++ subtree/vision-core/library/src/main/java/com/lakeel/altla/vision/data/repository/firebase/UserTextureFileRepositoryImpl.java
@@ -10,27 +10,30 @@ import com.google.firebase.storage.UploadTask;
 import com.lakeel.altla.rx.firebase.storage.RxFirebaseStorageTask;
 import com.lakeel.altla.rx.tasks.RxGmsTask;
 import com.lakeel.altla.vision.ArgumentNullException;
-import com.lakeel.altla.vision.domain.repository.TextureFileRepository;
+import com.lakeel.altla.vision.domain.repository.UserTextureFileRepository;
 
 import java.io.File;
 import java.io.InputStream;
 
 import rx.Single;
 
-public final class TextureFileRepositoryImpl implements TextureFileRepository {
+public final class UserTextureFileRepositoryImpl implements UserTextureFileRepository {
 
     private static final String PATH_USER_TEXTURES = "userTextures";
 
     private final StorageReference rootReference;
 
-    public TextureFileRepositoryImpl(StorageReference rootReference) {
+    public UserTextureFileRepositoryImpl(StorageReference rootReference) {
         if (rootReference == null) throw new ArgumentNullException("rootReference");
 
         this.rootReference = rootReference;
     }
 
     @Override
-    public Single<String> save(String fileId, InputStream stream, OnProgressListener onProgressListener) {
+    public Single<String> save(String id, InputStream stream, OnProgressListener onProgressListener) {
+        if (id == null) throw new ArgumentNullException("id");
+        if (stream == null) throw new ArgumentNullException("stream");
+
         // NOTE:
         //
         // Callbacks of the task for Firebase Storage are called by the thread for Firebase Storage.
@@ -39,34 +42,37 @@ public final class TextureFileRepositoryImpl implements TextureFileRepository {
 
         StorageReference reference = rootReference.child(PATH_USER_TEXTURES)
                                                   .child(resolveCurrentUserId())
-                                                  .child(fileId);
+                                                  .child(id);
         UploadTask task = reference.putStream(stream);
 
         return RxFirebaseStorageTask.asSingle(task, onProgressListener::onProgress)
-                                    .map(snapshot -> fileId);
+                                    .map(snapshot -> id);
     }
 
     @Override
-    public Single<String> delete(String fileId) {
+    public Single<String> delete(String id) {
+        if (id == null) throw new ArgumentNullException("id");
 
         StorageReference reference = rootReference.child(PATH_USER_TEXTURES)
                                                   .child(resolveCurrentUserId())
-                                                  .child(fileId);
+                                                  .child(id);
         Task<Void> task = reference.delete();
 
-        return RxGmsTask.asObservable(task).map(aVoid -> fileId).toSingle();
+        return RxGmsTask.asObservable(task).map(aVoid -> id).toSingle();
     }
 
     @Override
-    public Single<String> download(String fileId, File destination, OnProgressListener onProgressListener) {
+    public Single<String> download(String id, File destination, OnProgressListener onProgressListener) {
+        if (id == null) throw new ArgumentNullException("id");
+        if (destination == null) throw new ArgumentNullException("destination");
 
         StorageReference reference = rootReference.child(PATH_USER_TEXTURES)
                                                   .child(resolveCurrentUserId())
-                                                  .child(fileId);
+                                                  .child(id);
         FileDownloadTask task = reference.getFile(destination);
 
         return RxFirebaseStorageTask.asSingle(task, onProgressListener::onProgress)
-                                    .map(snapshot -> fileId);
+                                    .map(snapshot -> id);
     }
 
     private String resolveCurrentUserId() {
